# docs/api/openapi-config.yaml
openapi: 3.1.1
info:
  title: Blockasaurus config API
  description: Configuration management endpoints for Blockasaurus DNS proxy
  version: '1.0'
servers:
  - url: /api/config
paths:
...
  /upstream-groups/{name}/servers:
    get:
      operationId: listUpstreamServers
      tags: [upstream-groups]
      summary: List servers in an upstream group
      parameters:
        - $ref: '#/components/parameters/UpstreamGroupName'
      responses:
        '200':
          description: List of upstream servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UpstreamServer'
    post:
      operationId: createUpstreamServer
      tags: [upstream-groups]
      summary: Add a server to an upstream group
      parameters:
        - $ref: '#/components/parameters/UpstreamGroupName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpstreamServerInput'
      responses:
        '201':
          description: Created upstream server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpstreamServer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /upstream-groups/{name}/servers/{id}:
    put:
      operationId: updateUpstreamServer
      tags: [upstream-groups]
      summary: Update an upstream server
      parameters:
        - $ref: '#/components/parameters/UpstreamGroupName'
        - $ref: '#/components/parameters/ResourceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpstreamServerInput'
      responses:
        '200':
          description: Updated upstream server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpstreamServer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteUpstreamServer
      tags: [upstream-groups]
      summary: Delete an upstream server
      parameters:
        - $ref: '#/components/parameters/UpstreamGroupName'
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
...
    UpstreamGroupName:
      name: name
      in: path
      required: true
      schema:
        type: string
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    UpstreamServer:
      type: object
      required: [id, group_name, url, position, enabled]
      properties:
        id:
          type: integer
        group_name:
          type: string
        url:
          type: string
          description: >-
            Upstream server in `[net:]host[:port][/path][#commonName]` form, or a
            DNS stamp (`sdns://...`). Supported `net` prefixes are `tcp+udp`
            (default when omitted), `tcp-tls`, `https:` and `quic:`. For the
            latter two the `//` after the colon is optional, so `quic://dns.adguard.com`
            is accepted as well (AdGuard compatibility).
          example: quic:dns.adguard.com
        position:
          type: integer
        enabled:
          type: boolean
    UpstreamServerInput:
      type: object
      required: [url, enabled]
      properties:
        url:
          type: string
          description: >-
            Upstream server in `[net:]host[:port][/path][#commonName]` form, or a
            DNS stamp (`sdns://...`). Supported `net` prefixes are `tcp+udp`
            (default when omitted), `tcp-tls`, `https:` and `quic:`. For the
            latter two the `//` after the colon is optional, so `quic://dns.adguard.com`
            is accepted as well (AdGuard compatibility).
          example: quic:dns.adguard.com
        position:
          type: integer
        enabled:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string